WIP: export Google sheet to diffable form
Towards #24

This is the initial pass at an AppScript to turn a
Google Sheet into something that can be usefully
diffed in a git repo for human readable deltas
associated with changes.

To run, go to a Sheet->Extensions->Apps Script
Replace the script file in the editor with export.js
content and save it.

Select the onOpen function in the dropdown to
install handler into the sheet.
Accept the permissions prompts.

Go back to the Sheet->Custom Functions->Export

TODO:
* serialize comments
* serialize input data independent of sheet
* save to a subdir with the Sheet name to allow for
  multiple inflight sheet versions