Minor serialization fixes and improvements
"auto" is now dropped from serialization, same as "text".
It's not deleted in the same way text is because I'm not
100% of it's utility, but it shouldn't be adding noise to
the json.

Change of format for active LTG rate from 2dp % to 0dp

Moves the note field to be the last serialized as it's the
most length variable.

Fixes attempt to serialize formulas as dates when they're
formatted with iso-date. This was resulting in epoc dates
showing up in serialization for those cells.

On a different note, these issus were discovered while I was
doing a text diff based review of whether the change for #28
had been migrated from my personal sheet correctly.
This proves out the utility of the serialization mechanism,
although whether it was worth the opportunity cost at this
time is tbd.
After review I can declare the prior commits:
fix #28