Add Tweaks and clarify tax estimate
Added two placeholder tweaks:
* #10 - select election treatment
* #25 - share or lot display for value

Re-arranged the federal income+stg and ltg.
Added a total and made it clear the other two values
are the component parts of that. Previously the total
was on the std/income line and the ltg wasn't included
in the total summation at the bottom of it's column...
which is just confusing UX.