Foundation for per-lot cash/stock preference
Consolidated the following tweaks:
* AVGO share quantaties
* Cash consideration
* Election treatment

The first two were doing the same thing, but allowed
independent selection. Lack of consistency within a
transaction doesn't seem valid.

The last was a placeholder for the work these changes
lay the basis for. There are now three choices for the
ratio used to calculate cash/stock portions.

Pro-rata
This uses the ratio derived for the holding as a whole, using
the quantity of vmw converted to avgo vs vmw for cash.

Derived eTrade Per-lot
The number of shares per lot shown in eTrade OSPS holdings
are not consistent with a single ratio. This indicates that
eTrade is used a per-lot ratio that satisfies in aggregate
the global ratio for a given total holding.

Manual Per-lot
Allows manual selection on a per-lot basis of a preference
for cash or stock. This selection is used to fill a per-lot
ratio. It's hidden behind a dropdown because there's likely
only 3 values across the lots.
* as much cash as possible
* as much stock as possible
* a balancing ratio to get the holding as a whole to conform

Added additional columns to ESPP and RSU so we have one for
each choice for:
* cash received
* shares received
And an additional column for each of those to capture the
active value from current choice.

As part of this, I've pulled the active choice into a
reference cell to allow short names in SWITCHs.

Fixed various SUM references that incorrectly referenced row
3 as the starting row instead of 5. Was not causing numeric
errors but was technically incorrect.