Add totals to top of ESPP & RSU sheets
This adds an additional total row to the top of the datagrids.
This has the following benefits:

1. the value you mostly care about is visible without scrolling
2. the formulas that use totals now all use the same row, and
   aren't susceptible change if more rows are added in the grid
   making diffs less noisy

The specific prompt is so that I can easily see changes to
totals from changing cash/shares selections for specific lots.

Validation of all formulae updates being complete was done via
a regexp search for the RSU and ESPP total rows in the export.