Expanded on incorrect Supplement numbers
Found a note on Vanguard about there being no duty to provide basis numbers that are adjusted for filing.

That makes everything best effort, and in this case eTrade's best effort has failed us. Noted that the primary issue is _incorrect_ instead of _absent_ figures in the Supplement. No data is better than bad data.