55 import of v12 has issues with the use for fraction column (#56)
This promotes the import instructions from the latest release to its own section, and adds fixup information to work around #30.

Refines the instructions a little, and updates references to menu names for the v0.1.2 release

The import is a pain and also requires fixup. I've added a link to the Install section to let people go directly to the shared sheet.