Discussion forum enablement and documentation enhancements
Enables the Discussions feature for the repo. I've never used this, so don't know how useful it'll be. Added a link to it in the readme.

Where there are links to the latest shared sheet that is now via a link shortner to avoid failure to update on new release.

Added links to the specific usage steps in the collect necessary inputs list.

Broke out the BenefitsHistory download into a separate usage section.

Minor fixes in release section from prior releases.