License as Creative Commons Non-commercial 4.0
I had a non-commercial rider noted in the repo, but a user pointed out that's hard to notice and apply vs inline in the license. Given I've not yet had any contributors there's no issue with a fiat switch to the Creative Commons Non-commercial.

I'll keep the repo note about contacting me for dual-licensing for commercial use and add an explicit interpretation that use for filing and tracking by my (ex)colleagues is acceptable non-commercial use.