Make bottom column totals a simple mirror of upper
I had a frozen row with totals at the top of the RSU & ESPP datasheets.
I also had totals at the bottom because that's where people normally
look for them.
The bottom totals were SUMs as were the top. This meant more places to
fixup when adding rows or making other changes.

This change makes the bottom totals a pure mirror of the upper totals so
 that insertion of rows does not prompt a formula change.