Towards automated fractional lot
This calculates the standard deviation of the ratios from values in
AVGO qty from eTrade column. It constructs a threshold below which
a lot is highly likely to be the fractional lot.

A different approach to this may be to find the flat minimum ratio and
say that is likely the lot.

Either way, I need a means to associate this value with the fractional
column. Cross-sheet references don't seem to work in Conditional
Formatting.