Substantial revision to in-sheet and repo documentation
This moves the last of the in-line step instructions out of the sheet and replaces them
with higher level guidance and directions to the repo.

In addition to that, this adds a checklist to the Summary page to help people track
where they are in running through the sheet.

Finally this enriches the usage documentation for entering ESPP/RSU, including adding
rows when the prepopulated fields are absent for the lot.

Adds docs for entry of F8949 values into Turbotax

Fixes #53
Fixes #42
Fixes #88