Update ESPP imputed income calc with lesser of clause
When determining imputed income on which you have to pay tax (the bargain element) for qualified shares, there's a "lesser of" clause.

You pay either the lesser of the:
1. discount as of grant date (grant date FMV * discount percentage)
2. capital gain over actual purchase amount from eventual sale

This adds support for that clause. It's not useful as of writing because it'll never be active, but needed for completeness for tracking of future sales.

Fixes #15