Add colunms to ESPP for lot and pure post-merger basis
The post-merger basis column contained the imputed income from ESPP sale.
This adds a pure post-merger basis column suitable for provision into other
basis tracking systems, such as brokers or CPAs may maintain.

This adds a lot level basis column to match that in RSU so that we can present
the total pre-merger tax basis in summary. I don't know what this'll be useful
for but I needed it when figuring out the impact of the use of estimated RoC
value vs the finalized valued for 2021 distribution.

Adds the pre-merger tax basis to the summary.

Fixes #102