Correct 8949 column for non-covered security basis
Per the form instructions, for non-covered securities (ie. RSUs) where
the basis is not reported to the IRS, the corrected basis should go into
the regular basis column (e) and the adjustment (g) should contain
numeric zero.

The same gain results with either approach, but presumably there's a
reason for the difference in entry.

Experiments with TurboTax show that it put the corrected basis value
into the correct column for RSUs even if you enter it as a negative
adjustment.

Fixes #100