Update references to use upper SUM or full range
Nothing should be referencing rows below 5 in either of the RSU or ESPP
sheets.

For logic needing a SUM, it should be referencing row 5 of the column.
For logic needing to work on the raw range, it should be using
(Col7:Col) as the range to include the entire column.