RSU & ESPP lots use dynamic lookup of price
This is the primary commit shifting us to using VLOOKUP to get price info based
on date, instead of having people enter their share quantities into the "correct"
row.

This unbounds the lot rows in ESPP & RSU datasheets by removing the bottom
totals row, updating any ranges to refer to the entire column from row 7 down.

For ESPP the "default" number of lots (ie. rows with formula populated) is set to
allow one ESPP lot per 6 months back to 2007 when the current ESPP plan was
established.

For RSUs the "default" is 150 which is just a guestimate but should suffice for
pretty much everyone.

Additionally, in ESPP, this adjusts the the carry/contribution formula so that
it looks sane if the optional values for contribution aren't filled in.

The price information has been sourced from wligithub/tax-tool @ 93c2b2c2
with Wei's permission. See data/vmw-historical-price.csv
I validated the prices match with all price information I previously had
in the sheet.

Tries to improve in-sheet doc by adding Notes to the first entry row of all inputs.
Adds sub-headings for Required/Optional in-sheet as the colour coding does not
seem to have been effective as a UX indicator.

Fixes #9