Improve doc around entering AVGO qty from eTrade
This updates the documentation to reflect changes made to column names for disambiguation. It should have been done when the column names were updated but was missed.

This also updates the screenshot to reflect the latest sheet.

Removes reference to the trade history cvs as I can no longer find a way to download that from eTrade. I've instead pointed at the BenefitHistory sheet which seems to be consistently available if somewhat awkward to process.

Fixes #128