Clarify date input format
The date format was ISO, eg. 2001-01-31, but with that format for display
it's not apparent if the date has been entered incorrectly for RSUs where
the date tends to be on the first of the month.

As such this updates the format to 2021-Jan-31 so that it's utterly
unambiguous for read back.

Fixes #130
Fixes #129