Support case where ESPP lot is 100% cash
This includes the imputed income element if an ESPP lot was
converted to 100% during the merger.

This does not allow someone who chose 100% cash to use the
sheet without issue however, as the gain is still floored at 0 and
loss is suppressed.
It will suffice for ESPPs however as no lots incurred a loss...and
RSUs are very simple for the 100% cash case.