Set Form8949 display default to cash-only
People seem confused with the true economic value display. Given the
because the TurboTax and Form8949 do state to use the exact values
from the 1099, I've changed the default to cash-only.

I'll be following up with a CPA about the validity of differing proceeds
values because it's necessary for per-lot adjustment.