Add per-lot post-merger tax basis column
This adds a per-lot post-merger tax basis column to aid with reporting
when shares have been sold.

For ESPP the column is the basis adjusted by imputed income. The
primary reason for wanting a lot basis is due to sale in which case
the adjusted basis is the one to use.

Fixes #138